fixes Swift tools version
I'm not sure if this was working for other people, but we are still finding a conflict between Commander & Stencil.  When trying to resolve my dependencies using Swift tools version 4, this is the dependency graph I get:

└── Stencil<https://github.com/kylef/Stencil.git@0.10.1>
    ├── PathKit<https://github.com/kylef/PathKit.git@0.8.0>
    │   └── Spectre<https://github.com/kylef/Spectre.git@0.7.2>
    └── Spectre<https://github.com/kylef/Spectre.git@0.7.2>

Based on the documentation here: https://swift.org/blog/swift-package-manager-manifest-api-redesign/

> Since Swift 3 Package Manager doesn’t understand the Swift 4 manifest format, it will automatically ignore the Git tags that contain a Swift 4 manifest. So, if a package upgrades to Swift 4 manifest, Swift 3 Package Manager will pick the last tag which contains the Swift 3 manifest. However, the Package Manager in Swift 4 will pick the latest available version, regardless of manifest version.

It sounds like because Stencil's Package specifies tools version 3, the tags in Spectre and PathKit which are specifying version 4 are getting ignored, thus causing a conflict between Stencil and Commander.