Migrate away from PathKit to Foundation
Migrate from PathKit to Foundation for the file management operations.
This enables the use of Stencil on Windows.  With this set of changes,
the complete test suite passes on Windows.

This is a breaking change as it removes some deprecated APIs as well as
changes the shape of the API to fully remove the dependency on PathKit.